Abstract texture arrays
